Fixes issue where indexing numpy arrays with a float was allowed prior to NumPy 1.12; now explicitly casts a float to int.

+++ flyvbjerg_petersen_std_err.py
@@ -51,7 +51,7 @@ def block_transformation(series):
 
     """
     n_steps = series.size
-    n_steps_p = np.floor(n_steps/2.)
+    n_steps_p = np.floor(n_steps/2.).astype('int')
     output = 0.5 * (series[::2][:n_steps_p] + series[1::2][:n_steps_p])
     return output
